HTTP Proxy Support
http://stream-recorder.com/forum/compile-rtmpdump-windows-http-proxy-support-applying-t11027.html
http://lists.mplayerhq.hu/pipermail/rtmpdump/2010-August/001133.html

Here is a modified version of my patch which uses the http_proxy
environment variable instead of command line arguments to specify the
HTTP proxy.  This patch also includes extra checking for the HTTP
response code in HTTP_read().  This checking is required to prevent an
infinite loop when it has a value not equal to 200.